+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## Fixed
+
+- `/detectors/aws` no longer fails if instance metadata is not available (e.g. not running in AWS) (#401)
+
 ## [0.13.0] - 2020-10-09
 
 ## Added